Match recipe
I matched the recipe as I understood it.  I don't believe there are more events to respond to as the YouTube API does not have a "Skip" for example, which is too bad.  

I also moved "extensions" to the context object in a couple of places, instead of the results object.